+++ lib/flutter_curiosity.dart
@@ -7,6 +7,6 @@ export 'platform/desktop_window.dart';
 export 'platform/platform.dart';
 export 'platform/result_listener.dart';
 export 'tools/app_device.dart';
-export 'tools/gallery_tools.dart';
+export 'tools/gallery.dart';
 export 'tools/keyboard.dart';
 export 'tools/setting.dart';