+++ lib/tools/internal.dart
@@ -1,6 +1,7 @@
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_curiosity/constant/constant.dart';
 import 'package:flutter_curiosity/platform/platform.dart';
 
 final MediaQueryData mediaQueryData = MediaQueryData.fromWindow(window);
@@ -56,5 +57,78 @@ void _segmentationLog(String msg) {
         break;
       }
     }
+  }
+}
+
+String macOSSettingPathToString(MacOSSettingPath path) {
+  switch (path) {
+    case MacOSSettingPath.accessibilityMain:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess';
+    case MacOSSettingPath.accessibilityDisplay:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess?Seeing_Display';
+    case MacOSSettingPath.accessibilityZoom:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess?Seeing_Zoom';
+    case MacOSSettingPath.accessibilityVoiceOver:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess?Seeing_VoiceOver';
+    case MacOSSettingPath.accessibilityDescriptions:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess?Media_Descriptions';
+    case MacOSSettingPath.accessibilityCaptions:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess?Captioning';
+    case MacOSSettingPath.accessibilityAudio:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess?Hearing';
+    case MacOSSettingPath.accessibilityKeyboard:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess?Keyboard';
+    case MacOSSettingPath.accessibilityMouseTrackpad:
+      return 'x-apple.systempreferences:com.apple.preference.universalaccess?Mouse';
+    case MacOSSettingPath.securityMain:
+      return 'x-apple.systempreferences:com.apple.preference.security';
+    case MacOSSettingPath.securityGeneral:
+      return 'x-apple.systempreferences:com.apple.preference.security?General';
+    case MacOSSettingPath.securityFileVault:
+      return 'x-apple.systempreferences:com.apple.preference.security?FDE';
+    case MacOSSettingPath.securityFirewall:
+      return 'x-apple.systempreferences:com.apple.preference.security?Firewall';
+    case MacOSSettingPath.securityAdvanced:
+      return 'x-apple.systempreferences:com.apple.preference.security?Advanced';
+    case MacOSSettingPath.securityPrivacy:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy';
+    case MacOSSettingPath.securityPrivacyAccessibility:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy_Accessibility';
+    case MacOSSettingPath.securityPrivacyAssistive:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy_Assistive';
+    case MacOSSettingPath.securityPrivacyAllFiles:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy_AllFiles';
+    case MacOSSettingPath.securityPrivacyLocationServices:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy_LocationServices';
+    case MacOSSettingPath.securityPrivacyContacts:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy_Contacts';
+    case MacOSSettingPath.securityPrivacyDiagnosticsUsage:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy_Diagnostics';
+    case MacOSSettingPath.securityPrivacyCalendars:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy_Calendars';
+    case MacOSSettingPath.securityPrivacyReminders:
+      return 'x-apple.systempreferences:com.apple.preference.security?Privacy_Reminders';
+    case MacOSSettingPath.speechDictation:
+      return 'x-apple.systempreferences:com.apple.preference.speech?Dictation';
+    case MacOSSettingPath.speechTextToSpeech:
+      return 'x-apple.systempreferences:com.apple.preference.speech?TTS';
+    case MacOSSettingPath.sharingMain:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing';
+    case MacOSSettingPath.sharingScreenSharing:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing?Services_ScreenSharing';
+    case MacOSSettingPath.sharingFileSharing:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing?Services_PersonalFileSharing';
+    case MacOSSettingPath.sharingPrinterSharing:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing?Services_PrinterSharing';
+    case MacOSSettingPath.sharingRemoteLogin:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing?Services_RemoteLogin';
+    case MacOSSettingPath.sharingRemoteManagement:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing?Services_ARDService';
+    case MacOSSettingPath.sharingRemoteAppleEvents:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing?Services_RemoteAppleEvent';
+    case MacOSSettingPath.sharingInternetSharing:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing?Internet';
+    case MacOSSettingPath.sharingBluetoothSharing:
+      return 'x-apple.systempreferences:com.apple.preferences.sharing?Services_BluetoothSharing';
   }
 }